refactor: harden bff build and health checks
