+++ apps/bff/Dockerfile
@@ -15,7 +15,10 @@ RUN pnpm fetch --filter ./packages/sdk...
 RUN pnpm fetch --filter ./apps/bff...
 RUN pnpm fetch --prod --filter ./apps/bff...
 COPY . .
-RUN pnpm install --offline --filter ./apps/bff... --frozen-lockfile
+## Ensure dependencies for the SDK and BFF are installed prior to building
+RUN pnpm install --offline --frozen-lockfile \
+      --filter ./packages/sdk... \
+      --filter ./apps/bff...
 RUN pnpm --filter ./packages/sdk... build
 RUN pnpm --filter ./apps/bff... build
