--- apps/bff/src/health/health.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { BtcpayModule } from '../btcpay/btcpay.module';
-import { HealthController } from './health.controller';
-
-@Module({
-  imports: [BtcpayModule],
-  controllers: [HealthController]
-})
-export class HealthModule {}
